+++ changelog.md
@@ -1,8 +1,15 @@
 # 1.1.1
+### Additions
+- Added compatibility with Create
+  - Deployer (deploying) recipes for waxing/scraping silver blocks
+  - Spout (filling) recipes for oxidizing silver blocks (Similar to Create: Oxidized)
+- Added compatibility with EMI
+  - World Interaction recipes for waxing/scraping silver blocks
+
 ### Fixes
 - Added missing crafting recipes for Silver Bulb and its variants.
 - Added missing smelting recipes for Silver Ore and Deepslate Silver Ore blocks
-
+- Fix incorrect tag `ores_silver`. Changed to `ores/silver` to ensure compatibility with other mods.
 
 # 1.1.0
 ### Additions